Transfortation 완료

+++ Client/Assets/Scripts/Data/Data.Contents.cs
@@ -61,8 +61,8 @@ namespace Data
 	{
 		public int id;
 		public string regionName;
-		public float posX;
-		public float posY;
+		public int posX;
+		public int posY;
 	}
 
 	[Serializable]
